server/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
httpx
